refactor(viewer): delete unused point, viewer type
