+++ libs/insight-viewer/src/Annotation/components/AnnotationOverlay/utils/getIsComplexPolygon.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-plusplus */
 
 import { getIsIntersection } from './getIsIntersection'
-import { Point } from '../../../types'
+import { Point } from '../../../../types'
 
 /**
  * Check if polygon is a complex polygon with intersection
