+++ libs/insight-viewer/src/Annotation/components/AnnotationOverlay/utils/getIsIntersection.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import { Point } from '../../../../types'
 
 /**
  * Determine if ab and cd intersect
