+++ libs/insight-viewer/src/Annotation/components/TextDrawer/Typing.types.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../types'
+import { Point } from '../../../types'
 
 export interface TypingProps {
   points: Point[]
