+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/utils/getRulerConnectingLine.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import { Point } from '../../../../types'
 import { getRulerConnectingLine } from './getRulerConnectingLine'
 
 describe('getConnectingLinePoints: ', () => {
