+++ libs/insight-viewer/src/Annotation/utils/getAnnotationPointsAfterMoving.ts
@@ -2,7 +2,7 @@ import { getMovedPoints } from '../../utils/common/getMovedPoints'
 import { getCircleRadiusByCenter } from '../../utils/common/getCircleRadius'
 import { getCircleStartPoint } from './getCircleStartPoint'
 
-import type { Point } from '../types'
+import type { Point } from '../../types'
 
 export function getAreaAnnotationMovedPoints(prevPoints: Point[], movePoint: Point, currentPoint: Point) {
   const movedPoints = getMovedPoints({ prevPoints, editStartPoint: movePoint, currentPoint }) as [Point, Point]
